+++ run_developer_setup.sh
@@ -4,9 +4,9 @@
 pip install -r requirements.txt
 
 # Install in developer mode
-#pip install -e . --verbose
 
 # For some reason there is a bug with using pip and skbuild 
 # Calling setup.py directly seems to work though
 python setup.py clean
-python setup.py develop
+#python setup.py develop
+pip install -e .